Rubocop ready code
